set up @artsy/fresnel, restructure some components, remove ga by default
