+++ src/components/ui/Link.tsx
@@ -1,6 +1,4 @@
-/** @jsxRuntime classic */
-/** @jsx jsx */
-import { jsx } from "theme-ui"
+/** @jsxImportSource theme-ui */
 import NextLink from "next/link"
 
 interface Props {
