+++ src/components/ui/NavLink.tsx
@@ -1,6 +1,4 @@
-/** @jsxRuntime classic */
-/** @jsx jsx */
-import { jsx } from "theme-ui"
+/** @jsxImportSource theme-ui */
 import Link from "next/link"
 
 interface Props {
