+++ src/components/views/About.tsx
@@ -1,6 +1,6 @@
 import { MDXProvider } from "@mdx-js/react"
-import MarkdownContent from "../mdx/about.mdx"
-import Container from "../ui/Container"
+import MarkdownContent from "mdx/about.mdx"
+import Container from "components/ui/Container"
 
 const About: React.FC = () => {
   return (
